Fixed up all the warnings. Just some casting from pointer to unsigned integer for the printf's. Solves cases 23,24,25.
git-svn-id: https://gamescrafters.svn.sourceforge.net/svnroot/gamescrafters/GamesmanClassic/trunk@3267 a668be5d-ef11-cb4c-8bfd-e08cbb24f2ba

+++ gamesman/src/core/filedb/db.c
@@ -47,7 +47,7 @@ static gamesdb_frameid gamesdb_translate(gamesdb* db, gamesdb_pageid vpn) {
 	       
 		if (ppn->valid == TRUE) {
 			//this page got kicked out by n-chance
-            printf("translate: replaced = %llu\n", ppn);
+            printf("translate: replaced = %u\n", (unsigned int)ppn);
             assert(ppn->tag != vpn);
 			if (ppn->tag != vpn) {
 				//buffer page is valid but not the one we want
@@ -66,7 +66,7 @@ static gamesdb_frameid gamesdb_translate(gamesdb* db, gamesdb_pageid vpn) {
 	}
 	
 	if (DEBUG) {
-		printf("translate: vpn = %llu, ppn = %d tag = %llu\n", vpn, ppn, ppn->tag);
+		printf("translate: vpn = %llu, ppn = %u tag = %llu\n", vpn, (unsigned int)ppn, ppn->tag);
 	}
 	
     assert (ppn->tag == vpn);
